requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
joblib>=1.0.0
yfinance>=0.2.0
textblob>=0.15.3
fredapi>=0.5.0
python-dotenv>=0.19.0
requests>=2.26.0
nltk>=3.6.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8
transformers>=4.15.0
torch>=1.10.0
ta>=0.10.0 